Use JSClientChannel
